+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to
 - template: add clang.mk, pytorch.mk, runprog.mk;
 - pytorch: add compile script;
 #### Changed
+- Move template/targets/libso.sh to scripts/libso-multiver.sh;
 - Move vfs into fs directory;
 - Move alternative directory into insn directory;
 - Move aio directory into glibc;